calls: record the dynamic argument count in full (u8 frame slot saturated at 255)
CallFrame.argc and JSGenerator.argc were u8, and every dynamic
frame-push site clamped with @min(args.len, maxInt(u8)) — so
f.apply(null, args300), f(...args300), and Reflect.apply reported
arguments.length === 255 (§10.4.4 CreateUnmappedArgumentsObject
stamps the actual count) and truncated rest parameters (§15.2.4).
Generator and async frames additionally capped their saved register
file at 255 slots, dropping argument *values* past index 255 across
suspension — those registers are what a resumed body's arguments /
rest materialisers read.

Widen both counts to u32, size generator register files by the real
arg count, and widen the u8 loop indices in the §10.4.4 arguments-
object builder and the §15.2.4 rest collector. The direct-call
opcodes' u8 argc operand is untouched (a literal >255-arg call
exhausts the 256-register file first — reserveTemp errors out);
Bistromath reads argc only from those static operands and its u64
helper ABI, so the JIT tiers are unchanged.

Unit tests (red first): 300-arg apply → arguments.length and
arguments[299], rest via apply, spread call, a Math.max.apply
guard, and a generator observing count + values across resume.
test262 full sweep unchanged at 45241/4567 (90.83%); no fixture
exercises >255 args — gap-log entry follows.

+++ src/runtime/generator.zig
@@ -121,7 +121,10 @@ pub const JSGenerator = struct {
     this_value: Value = Value.undefined_,
     home_object: ?*JSObject = null,
     home_function: ?*@import("function.zig").JSFunction = null,
-    argc: u8 = 0,
+    /// Caller-supplied argument count, mirroring `CallFrame.argc`
+    /// (u32 — apply / spread counts are dynamic and must not
+    /// saturate; §10.4.4 `arguments.length` reads it on resume).
+    argc: u32 = 0,
     state: GeneratorState = .initial,
     /// Mark color. `gen.mark_color == heap.live_color` means "live
     /// this cycle". See `JSObject.mark_color` for the protocol.
@@ -249,7 +252,7 @@ pub const JSGenerator = struct {
     pub fn init(
         allocator: std.mem.Allocator,
         chunk: *const Chunk,
-        register_count: u8,
+        register_count: u32,
         captured_env: ?*Environment,
         this_value: Value,
     ) !*JSGenerator {
